added kdtree support and example
using the same approach as in the rest of the wrapper, wrapped
pcl::KdTreeFLANN from 'pcl/kdtree/kdtree_flann.h'. As in other cases,
appended a make_kdtree_flann function to the PointCloud class.
At this stage, the only method exposed is the nearestKSearch method,
in particular the version which takes a PointCloud and an index into
it to choose the point for comparision:
http://bit.ly/WWI9JS

Further added a convienience method which finds the k nearest
distances for all points in the provided point cloud. For an example,
see examples/kdtree.py